minor changes during dev
git-svn-id: svn://orca.physics.unc.edu/Orca@2562 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IpeV4 SLT/ORIpeV4SLTModel.h
@@ -282,8 +282,8 @@
 - (void)		  printInterrupt:(int)regIndex;
 //- (void)		  dumpTriggerRAM:(int)aPageIndex;
 
-- (void)		  writeReg:(unsigned short)index value:(unsigned long)aValue;
-- (unsigned long) readReg:(unsigned short) index;
+- (void)		  writeReg:(int)index value:(unsigned long)aValue;
+- (unsigned long) readReg:(int) index;
 - (unsigned long) readHwVersion;
 - (unsigned long long) readDeadTime;
 - (unsigned long long) readVetoTime;
